+++ lib/op.dart
@@ -1,2 +1,3 @@
-export 'src/binary/addable.dart' show Addable;
-export 'src/binary/multipliable.dart' show Multipliable;
+export 'src/binary/addable.dart' show Addable, AddableInternally;
+export 'src/binary/multipliable.dart' show Multipliable, MultipliableInternally;
+export 'src/unary/negatable.dart' show Negatable;