+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Added interface `Negatable<X>`
+- Added type aliases `AddableInternally<X>` and `MultipliableInternally<X>`
+
 ## 0.1.0
 
 - Initial development release